backend/simple_services/notification/requirements.txt
python-dotenv==1.0.1
requests==2.26.0
pyTelegramBotAPI==4.16.1
pika==1.3.2
Flask==3.0.2
SQLAlchemy==2.0.27
